project: updating from local disk.

+++ src/main/java/pro/techdict/bib/bibserver/daos/UserDetailsRepository.java
@@ -3,4 +3,4 @@ package pro.techdict.bib.bibserver.daos;
 import org.springframework.data.repository.CrudRepository;
 import pro.techdict.bib.bibserver.entities.UserDetails;
 
-public interface UserDetailsRepository extends CrudRepository<UserDetails, Integer> { }
+public interface UserDetailsRepository extends CrudRepository<UserDetails, Long> { }
